Create update_gsheet_data_v1.js
Replacement for the IMPORTRANGE function in Google Sheets which has limitations.

So this uses Apps Script to import all the data from the V/F Summary pages of the Master Inspection Data and copy to the Report Generation Sheet. This data can then be used to filter and generate the reports as required.

This Apps Script Function will be saved separately as an individual file names "Update Data.gs" separate from the main report generation file.